% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{combine_slot}
\alias{combine_slot}
\title{combine_slot}
\usage{
combine_slot(x, y)
}
\arguments{
\item{x}{slot 1}

\item{y}{slot 2}
}
\value{
S4:combined_slot
}
\description{
Combine different slots together from same data for S4 object
}
