% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{compute_score_eachgroup}
\alias{compute_score_eachgroup}
\title{process function after modules have been defined}
\usage{
compute_score_eachgroup(
  cut_sle,
  group,
  diffgenes,
  hig_gene,
  sle_tab,
  cv_all,
  sd_all,
  mean_all,
  minModule,
  maxModule,
  size_effect
)
}
\arguments{
\item{cut_sle}{module information table (cluster, which module each gene belongs to)}

\item{group}{group (assay_name)}

\item{diffgenes}{same to DNBcompute}

\item{hig_gene}{same to DNBcompute}

\item{sle_tab}{selected table}

\item{cv_all}{all cv values}

\item{sd_all}{all sd values}

\item{mean_all}{all mean values}

\item{minModule}{same to DNBcompute}

\item{maxModule}{same to DNBcompute}

\item{size_effect}{same to DNBcompute}
}
\description{
the parameters come from inner DNBcompute
}
