% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.example}
\alias{data.example}
\title{data.example}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 200 rows and 200 columns.
}
\usage{
data.example
}
\description{
data.example
}
\examples{
head(
 data.example
)
}
\keyword{datasets}
