% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{deltaPCC.test}
\alias{deltaPCC.test}
\title{statistical test for delta PCC}
\usage{
deltaPCC.test(delta, r0, n0, two_sides = TRUE)
}
\arguments{
\item{delta}{delta PCC}

\item{r0}{reference PCC}

\item{n0}{reference size}

\item{two_sides}{test of two-sides (default) or greater?}
}
\value{
p_value
}
\description{
statistical test for delta PCC
}
