% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diffgenes.example}
\alias{diffgenes.example}
\title{diffgenes.example}
\format{
An object of class \code{character} of length 50.
}
\usage{
diffgenes.example
}
\description{
diffgenes.example
}
\examples{
head(
 diffgenes.example
)
}
\keyword{datasets}
