% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\docType{class}
\name{DNB_module-class}
\alias{DNB_module-class}
\alias{mynew_module}
\title{define S4 object of DNB_module}
\value{
S4 object, DNB_module
}
\description{
define S4 object of DNB_module
}
\section{Slots}{

\describe{
\item{\code{MODULE}}{list.}

\item{\code{bestMODULE}}{logical.}
}}

