% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\docType{class}
\name{DNB_obj-class}
\alias{DNB_obj-class}
\alias{mynew_dnb}
\title{define S4 object of DNB_obj, includes the raw data(matrix or data.frame), the pre-results (S4-DNB_res) and the results of filtered (ntop max) modules (S4-DNB_res)}
\value{
S4 object, DNB_obj
}
\description{
define S4 object of DNB_obj, includes the raw data(matrix or data.frame), the pre-results (S4-DNB_res) and the results of filtered (ntop max) modules (S4-DNB_res)
}
\section{Slots}{

\describe{
\item{\code{data}}{ANY.}

\item{\code{pre_result}}{DNB_res.}

\item{\code{result}}{DNB_res.}
}}

