% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\docType{class}
\name{DNB_res-class}
\alias{DNB_res-class}
\alias{mynew_res}
\title{define S4 object of DNB_pre, includes the numeric results of each DNB-computation and S4-DNB_module}
\value{
S4 object, DNB_res
}
\description{
define S4 object of DNB_pre, includes the numeric results of each DNB-computation and S4-DNB_module
}
\section{Slots}{

\describe{
\item{\code{MEAN}}{numeric.}

\item{\code{SD}}{numeric.}

\item{\code{CV}}{numeric.}

\item{\code{MODULEs}}{DNB_module.}

\item{\code{PCC_IN}}{numeric.}

\item{\code{PCC_OUT}}{numeric.}

\item{\code{SCORE}}{numeric.}

\item{\code{rank}}{numeric.}

\item{\code{resource}}{character.}

\item{\code{rank_all}}{numeric.}
}}

