% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{DNBcompute_custom.default}
\alias{DNBcompute_custom.default}
\title{Compute the Dynamic Network Biomarkers(DNB) model with customized Modules}
\usage{
\method{DNBcompute_custom}{default}(
  data,
  module_list,
  meta,
  meta_levels = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{data}{the gene expression matrix,
which can be a single-cell RNA-seq GEM with at least three group/clusters
or a matrix merging bulk GEMs from at least three different sample}

\item{module_list}{a customized list of module gene}

\item{meta}{a data.frame with rownames as cell-id as well as one column of group infomation}

\item{meta_levels}{the order of meta group, default ordered by decreasing if NULL}

\item{quiet}{do not message}

\item{...}{not use}
}
\value{
S3:DNB_output
}
\description{
Compute the Dynamic Network Biomarkers(DNB) model with customized Modules
}
