% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DNBcompute_custom}
\alias{DNBcompute_custom}
\title{DNBcompute_custom}
\usage{
DNBcompute_custom(
  data,
  module_list,
  meta = NULL,
  meta_levels = NULL,
  force_allgene = TRUE,
  size_effect = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{data}{input data, S4:DNB_output or gene expression matrix}

\item{module_list}{a customized list of module gene}

\item{meta}{a data.frame with rownames as cell-id as well as one column of group infomation, 
use if data is gene expression matrix}

\item{meta_levels}{the order of meta group, default ordered by decreasing if NULL, 
use if data is gene expression matrix}

\item{force_allgene}{whether force to use all genes from data, default TRUE}

\item{size_effect}{whether consider the effect of sample size when compute CI of DNB, default TRUE}

\item{quiet}{do not message}

\item{...}{for future use}
}
\value{
S3:DNB_output
}
\description{
Compute the Dynamic Network Biomarkers(DNB) model with customized Modules
}
\details{
input data from existing DNB_output object, or new data and meta

input customized modules from a list of gene set(s)

return a S3 object includes several S4 objects
}
\examples{
data(data.example)
data(meta.example)
data(module_list.example)
b <- DNBcompute_custom(data.example, module_list.example, meta.example)
b
DNBplot(b, group = "USER_CUSTOMIZED", ranking = 1, show = TRUE, save_pdf = FALSE)
DNBplot(b, group = "USER_CUSTOMIZED", ranking = 1, show = TRUE, save_pdf = FALSE)

}
\author{
Kaiyu Wang, in ChenLab of CAS, Shanghai, China
}
