% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{DNBcomputeSingle,DNB_obj-method}
\alias{DNBcomputeSingle,DNB_obj-method}
\title{DNBcomputeSingle for DNB_obj}
\usage{
\S4method{DNBcomputeSingle}{DNB_obj}(object, module_list)
}
\arguments{
\item{object}{S4:DNB_obj}

\item{module_list}{a list of module genes}
}
\description{
DNBcomputeSingle for DNB_obj
}
