% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{DNBcomputeSingle}
\alias{DNBcomputeSingle}
\title{Compute data of single group for all modules from pre-result, and add the result into slot result}
\usage{
DNBcomputeSingle(object, ...)
}
\arguments{
\item{object}{S4:DNB_obj}

\item{...}{parameter: module_list, a list of modules, each with a gene vector, and others}
}
\value{
S4:DNB_obj
}
\description{
Compute data of single group for all modules from pre-result, and add the result into slot result
}
