% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{DNBcomputeSingle_DNB_obj}
\alias{DNBcomputeSingle_DNB_obj}
\title{DNBcomputeSingle for DNB_obj}
\usage{
DNBcomputeSingle_DNB_obj(object, module_list)
}
\arguments{
\item{object}{S4:DNB_obj}

\item{module_list}{a list of module genes}
}
\value{
DNB_obj
}
\description{
DNBcomputeSingle for DNB_obj
}
