% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DNBfilter}
\alias{DNBfilter}
\title{DNBfilter}
\usage{
DNBfilter(
  DNB_output,
  ntop,
  force_allgene = FALSE,
  size_effect = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{DNB_output}{S3:DNB_output, from output of DNBcompute}

\item{ntop}{the numbers of modules to filter}

\item{force_allgene}{whether force to use all genes from data, default FALSE (use the gene sets from output of DNBcompute)}

\item{size_effect}{whether consider the effect of sample size when compute CI of DNB, default TRUE}

\item{quiet}{do not message}
}
\value{
S3:DNB_output
}
\description{
Filter DNB model by selecting the most N modules from each group with max score,

and apply these modules to all group to recompute DNB model
}
\examples{
data(data.example)
data(meta.example)
a <- DNBcompute(data.example, meta.example)
b <- DNBfilter(a, ntop = 5)
b

}
\author{
Kaiyu Wang, in ChenLab of CAS, Shanghai, China
}
