% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{DNBplot.data.frame}
\alias{DNBplot.data.frame}
\title{DNBplot for data.frame}
\usage{
\method{DNBplot}{data.frame}(
  object,
  ranking = NULL,
  group = NULL,
  show = TRUE,
  save_pdf = FALSE,
  file_prefix = NULL,
  ...
)
}
\arguments{
\item{object}{data.frame of scores}

\item{ranking}{not use}

\item{group}{not use}

\item{show}{whether to show}

\item{save_pdf}{whether to save pdf file}

\item{file_prefix}{the file prefix if save_pdf}

\item{...}{not use}
}
\description{
DNBplot for data.frame
}
