% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DNBplot}
\alias{DNBplot}
\title{DNBplot}
\usage{
DNBplot(
  object,
  ranking = NULL,
  group = NULL,
  resource = NULL,
  show = TRUE,
  save_pdf = FALSE,
  file_prefix = NULL,
  meta_levels = NULL,
  ...
)
}
\arguments{
\item{object}{S3:DNB_output or df_score (output of ScoreExtract)}

\item{ranking}{for DNB_output, default 1 if NULL}

\item{group}{for DNB_output, default random selected if NULL}

\item{resource}{the actual module name, ranking & group will be ignored if use}

\item{show}{whether to show, default TRUE}

\item{save_pdf}{whether to save pdf file, default FALSE}

\item{file_prefix}{the file prefix if save_pdf, default NULL}

\item{meta_levels}{the order of meta-group in the plots, default levels(df_score$Names) if NULL}

\item{...}{for future use}
}
\value{
plot or pdf
}
\description{
Visualize the DNB score information
}
\examples{
data(data.example)
data(meta.example)
a <- DNBcompute(data.example, meta.example)
b <- DNBfilter(a, ntop = 5)
DNBplot(b, show = TRUE, save_pdf = FALSE)

}
\author{
Kaiyu Wang, in ChenLab of CAS, Shanghai, China
}
