% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{flattenCorrMatrix}
\alias{flattenCorrMatrix}
\title{flattenCorrMatrix}
\usage{
flattenCorrMatrix(cor_mtx)
}
\arguments{
\item{cor_mtx}{correlation matrix}
}
\value{
a data.frame including network nodes and edges (feature1/2, cor)
}
\description{
convert correlation matrix into data.frame

if the element (edge) is NA, this feature couple (nodes) will be ignored
}
