% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{getGlobalScore}
\alias{getGlobalScore}
\title{getGlobalScore}
\usage{
getGlobalScore(object, meta_name = NULL, quiet = FALSE)
}
\arguments{
\item{object}{S4::SSN}

\item{meta_name}{the name of meta.data (in colnames of object@score@meta.data)}

\item{quiet}{do not warn when meta_name is NULL}
}
\value{
GlobalScore matrix
}
\description{
get exact GlobalScore matrix from SSN object
}
