% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getMaxRanking.DNB_output}
\alias{getMaxRanking.DNB_output}
\title{Get the max Ranking of exactly group in S3:DNB_output after function DNBfilter}
\usage{
\method{getMaxRanking}{DNB_output}(object, group, ...)
}
\arguments{
\item{object}{S3:DNB_output}

\item{group}{which sub-group of S3-DNB_output to extract}

\item{...}{not use}
}
\value{
numeric value
}
\description{
Get the max Ranking of exactly group in S3:DNB_output after function DNBfilter
}
