% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getModuleGenes.DNB_obj}
\alias{getModuleGenes.DNB_obj}
\title{Get the gene list of exactly module in S4:DNB_obj}
\usage{
\method{getModuleGenes}{DNB_obj}(object, resource, ...)
}
\arguments{
\item{object}{S4:DNB_obj}

\item{resource}{that is, the actual module name}

\item{...}{not use}
}
\value{
vector of genes or NULL
}
\description{
Get the gene list of exactly module in S4:DNB_obj
}
