% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getModuleGenes.DNB_output}
\alias{getModuleGenes.DNB_output}
\title{Get the gene list of exactly module in S3:DNB_output}
\usage{
\method{getModuleGenes}{DNB_output}(object, resource, ...)
}
\arguments{
\item{object}{S3:DNB_output}

\item{resource}{that is, the actual module name}

\item{...}{not use}
}
\value{
vector of genes or character()
}
\description{
Get the gene list of exactly module in S3:DNB_output
}
