% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getModuleGenes}
\alias{getModuleGenes}
\title{getModuleGenes}
\usage{
getModuleGenes(object, resource, ...)
}
\arguments{
\item{object}{S3:DNB_output}

\item{resource}{that is, the actual module name}

\item{...}{for future use}
}
\value{
vector of genes
}
\description{
Get the gene list of exactly module in S3:DNB_output
}
\examples{
data(data.example)
data(meta.example)
a <- DNBcompute(data.example, meta.example)
b <- DNBfilter(a, ntop = 5)
A_11_genelist <- getModuleGenes(b, resource = "A_11")
print(A_11_genelist)

}
\author{
Kaiyu Wang, in ChenLab of CAS, Shanghai, China
}
