% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.example}
\alias{meta.example}
\title{meta.example}
\format{
An object of class \code{character} of length 200.
}
\usage{
meta.example
}
\description{
meta.example
}
\examples{
head(
 meta.example
)
}
\keyword{datasets}
