% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{module_list.example}
\alias{module_list.example}
\title{module_list.example}
\format{
An object of class \code{list} of length 2.
}
\usage{
module_list.example
}
\description{
module_list.example
}
\examples{
head(
 module_list.example
)
}
\keyword{datasets}
