% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mycat}
\alias{mycat}
\title{to print the process information}
\usage{
mycat(..., quiet = FALSE)
}
\arguments{
\item{...}{parameters of function cat}

\item{quiet}{not execute function cat}
}
\description{
to print the process information
}
