% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mycor}
\alias{mycor}
\title{to ignore the warning and error of correlation computation}
\usage{
mycor(x, y = NULL, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{other parameters of function cor}
}
\description{
to ignore the warning and error of correlation computation
}
