% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{mynew.DNB_output}
\alias{mynew.DNB_output}
\title{generate the S3-DNB_output or change the attribute of a list to S3-DNB_output}
\usage{
mynew.DNB_output(group, list_obj = NULL)
}
\arguments{
\item{group}{group information from meta}

\item{list_obj}{if assigned, transfer this list into S3-DNB_output by changing its attribute}
}
\value{
S3-DNB_output
}
\description{
generate the S3-DNB_output or change the attribute of a list to S3-DNB_output
}
