% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{mynew.DNB_res}
\alias{mynew.DNB_res}
\title{initialize or generate the S4-DNB_res}
\usage{
mynew.DNB_res(ntop = NULL, ...)
}
\arguments{
\item{ntop}{the number of sample, use when initialize}

\item{...}{parameters transmitting to function mynew_res() to generate S4-DNB_res after result has been computed}
}
\value{
S4 object, DNB_res
}
\description{
initialize or generate the S4-DNB_res
}
