% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{myprocess}
\alias{myprocess}
\title{Process function of each group data}
\usage{
myprocess(
  data,
  assay_name,
  diffgenes,
  allgenes,
  high_method,
  high_cutoff,
  cutree_method,
  cutree_cutoff,
  minModule,
  maxModule,
  quiet,
  fastMode,
  cluster_fun,
  cluster_args
)
}
\arguments{
\item{data}{same to DNBcompute}

\item{assay_name}{group (one of meta_levels)}

\item{diffgenes}{same to DNBcompute}

\item{allgenes}{same to DNBcompute}

\item{high_method}{same to DNBcompute}

\item{high_cutoff}{same to DNBcompute}

\item{cutree_method}{same to DNBcompute}

\item{cutree_cutoff}{same to DNBcompute}

\item{minModule}{same to DNBcompute}

\item{maxModule}{same to DNBcompute}

\item{quiet}{same to DNBcompute}

\item{fastMode}{same to DNBcompute}

\item{cluster_fun}{same to DNBcompute}

\item{cluster_args}{same to DNBcompute}
}
\description{
the parameters come from inner DNBcompute
}
