% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{NetworkCreate}
\alias{NetworkCreate}
\title{NetworkCreate}
\usage{
NetworkCreate(data, method = c("pearson", "spearman"), p_require = FALSE)
}
\arguments{
\item{data}{gene expression matrix (or data.frame), with col-sample and row-gene}

\item{method}{correlation method, pearson (PCC, default) or spearman}

\item{p_require}{whether to test and compute p_value?}
}
\value{
a list including matrix r and/or P
}
\description{
Compute genes network by PCC
}
