% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{PCCsum}
\alias{PCCsum}
\title{PCCsum}
\usage{
PCCsum(feature, network, ignore_idx = NULL)
}
\arguments{
\item{feature}{feature name (node)}

\item{network}{network data.frame (after flattenCorrMatrix())}

\item{ignore_idx}{index where it should be ignored (use when finding second-neighbors)}
}
\value{
a list including pcc and indices
}
\description{
get all PCC values of exact feature from feature-coupled network

if the element (edge) is NA, this feature couple (nodes) will be ignored
}
