% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{refNetworkCreate}
\alias{refNetworkCreate}
\title{refNetworkCreate}
\usage{
refNetworkCreate(
  ref_gex,
  method = "pearson",
  p_thre = 0.001,
  prior_network = NULL
)
}
\arguments{
\item{ref_gex}{gene expression matrix of reference}

\item{method}{correlation method, pearson (PCC, default) or spearman}

\item{p_thre}{p_value threshold for correlation}

\item{prior_network}{the prior network, such as experimental protein-protein interaction (PPI network)
NULL (default) or a data.frame that has 2 columns of node couples}
}
\value{
a list including matrix r, P and r_sig
}
\description{
Compute genes network from reference samples
}
