% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{resultAllExtract}
\alias{resultAllExtract}
\title{resultAllExtract}
\usage{
resultAllExtract(object, group, slot, ...)
}
\arguments{
\item{object}{the S3-DNB_output}

\item{group}{which sub-group of S3-DNB_output to extract}

\item{slot}{which S4-DNB_obj to extract, pre_result or result}

\item{...}{for future use}
}
\value{
data.frame
}
\description{
Get the whole result of slot pre_result or result from S4-DNB_obj and transfer that into a data.frame (matrix)
}
\examples{
data(data.example)
data(meta.example)
data(diffgenes.example)
a <- DNBcompute(data.example, meta.example, diffgenes.example)
b <- DNBfilter(a, ntop = 5)
df_score <- resultAllExtract(
    object = b,
    group = "A",
    slot = "pre_result"
)
df_score

}
