% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{ScoreExtract.DNB_output}
\alias{ScoreExtract.DNB_output}
\title{Extract the score information of @result from object (S3:DNB_output)}
\usage{
\method{ScoreExtract}{DNB_output}(object, ranking = NULL, group = NULL, resource = NULL, ...)
}
\arguments{
\item{object}{S3:DNB_output}

\item{ranking}{the ranking of exactly module, default 1 if NULL}

\item{group}{which group to select module, default random selected if NULL}

\item{resource}{the actual module name, ranking & group will be ignored if use}

\item{...}{not use}
}
\value{
score data.frame
}
\description{
Extract the score information of @result from object (S3:DNB_output)
}
