% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ScoreExtract}
\alias{ScoreExtract}
\title{ScoreExtract}
\usage{
ScoreExtract(
  object,
  ranking = NULL,
  group = NULL,
  resource = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{object}{S3:DNB_output}

\item{ranking}{the ranking of exactly module, default 1 if NULL}

\item{group}{which group to select module, default random selected if NULL}

\item{resource}{the actual module name, ranking & group will be ignored if use}

\item{quiet}{do not message}

\item{...}{for future use}
}
\value{
data.frame
}
\description{
Extract the score information from object (S3:DNB_output)
}
\examples{
data(data.example)
data(meta.example)
a <- DNBcompute(data.example, meta.example)
b <- DNBfilter(a, ntop = 5)
df_score <- ScoreExtract(
    object = b,
    ranking = NULL,
    group = NULL
)
df_score

}
\author{
Kaiyu Wang, in ChenLab of CAS, Shanghai, China
}
