% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{sED}
\alias{sED}
\title{sample Expression Deviation}
\usage{
sED(Xi, Xref, scale = TRUE)
}
\arguments{
\item{Xi}{gene expression from single-sample}

\item{Xref}{gene expression from reference-samples}

\item{scale}{whether to scale sED by the standard deviation of reference expression}
}
\description{
sample Expression Deviation
}
