% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{selectK}
\alias{selectK}
\title{selectK}
\usage{
selectK(object, K)
}
\arguments{
\item{object}{S4::SSN}

\item{K}{the top K genes by order of CI}
}
\value{
matrix with element 1/0 (TRUE/FALSE), colnames samples and rownames genes
}
\description{
select top K genes for every sample sorted by CI values
}
