% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{show,DNB_module-method}
\alias{show,DNB_module-method}
\title{show DNB_module}
\usage{
\S4method{show}{DNB_module}(object)
}
\arguments{
\item{object}{DNB_module}
}
\description{
show DNB_module
}
