% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{singleDNB}
\alias{singleDNB}
\title{Compute data for the modules list, and return score matrix}
\usage{
singleDNB(data, module_list)
}
\arguments{
\item{data}{data}

\item{module_list}{modules list}
}
\value{
score matrix
}
\description{
Compute data for the modules list, and return score matrix
}
