% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\docType{class}
\name{SSN-class}
\alias{SSN-class}
\alias{mynew_ssn}
\title{define S4 object of SSN}
\value{
S4 object, SSN
}
\description{
define S4 object of SSN
}
\section{Slots}{

\describe{
\item{\code{network}}{SSNnetwork}

\item{\code{score}}{SSNscore}
}}

