% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{SSNcompute}
\alias{SSNcompute}
\title{SSNcompute}
\usage{
SSNcompute(
  ss_gex,
  ref_gex,
  delta_p = 0.1,
  ref_p = 0.05,
  prior_network = NULL,
  nFirstOrderNeighbor = 3,
  nSecondOrderNeighbor = 1,
  sED_scale = TRUE,
  save_dir = NULL,
  quiet = FALSE
)
}
\arguments{
\item{ss_gex}{gene expression matrix of single sample(s)}

\item{ref_gex}{gene expression matrix of reference}

\item{delta_p}{p_value threshold for delta PCC test}

\item{ref_p}{p_value threshold for correlation test, when building gene network from reference}

\item{prior_network}{the prior network, such as experimental protein-protein interaction (PPI network)
NULL (default) or a data.frame that has 2 columns of node couples}

\item{nFirstOrderNeighbor}{the least number of first-order neighbor(s) for each gene, default 3}

\item{nSecondOrderNeighbor}{the least number of second-order neighbor(s) for each gene, default 1}

\item{sED_scale}{whether to scale sED by standard deviation}

\item{save_dir}{the directort for saving SSN network data.frame, default NULL (not save)}

\item{quiet}{do not print output of process during calculation (against verbose), default FALSE}
}
\value{
S4::SSN object, output will be stored in object@network
}
\description{
get all PCC values of exact feature from feature-coupled network

if the element (edge) is NA, this feature couple (nodes) will be ignored
}
