% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\docType{class}
\name{SSNnetwork-class}
\alias{SSNnetwork-class}
\alias{mynew_network}
\title{define S4 object of SSNnetwork}
\value{
S4 object, SSNnetwork
}
\description{
define S4 object of SSNnetwork
}
\section{Slots}{

\describe{
\item{\code{module}}{matrix}

\item{\code{network}}{list}

\item{\code{scores}}{list}
}}

