% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\docType{class}
\name{SSNscore-class}
\alias{SSNscore-class}
\alias{mynew_ssnscore}
\title{define S4 object of SSNscore}
\value{
S4 object, SSNscore
}
\description{
define S4 object of SSNscore
}
\section{Slots}{

\describe{
\item{\code{meta}}{character}

\item{\code{K}}{numeric}

\item{\code{freq}}{numeric}

\item{\code{name}}{character}

\item{\code{local_module}}{list}

\item{\code{global_score}}{data.frame}
}}

