% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{SSNscore_default}
\alias{SSNscore_default}
\title{SSNscore_default}
\usage{
SSNscore_default(
  object,
  K,
  freq = 0.5,
  meta = NULL,
  method = c("mean", "sum"),
  mean_by = c("K", "genes")
)
}
\arguments{
\item{object}{S4::SSN}

\item{K}{an integer, the top K genes by order of CI}

\item{freq}{percentage for selection of each group's genes.
That is, n_sample = freq * #repeats in one group,
and n_gene = (in this group how many repeats enriched this gene).
If n_gene >= n_sample, this gene is selected for this group}

\item{meta}{group information for all samples (repeats), 
a data.frame with 1-col, or NULL (default).
If NULL, object@score@meta.data should have any column and the first will be used}

\item{method}{method for aggregating the gene expression to calculate the score of each sample,
by average (mean, default) or summation (sum)}

\item{mean_by}{if method == mean, where the total number is K (default) or the actual number of genes}
}
\value{
S4::SSN object, result will be stored in object@score
}
\description{
compute the aggregated SSN score for each group
}
