% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\docType{class}
\name{SSNscores-class}
\alias{SSNscores-class}
\alias{mynew_ssnscores}
\title{define S4 object of SSNscores}
\value{
S4 object, SSNscores
}
\description{
define S4 object of SSNscores
}
\section{Slots}{

\describe{
\item{\code{meta.data}}{data.frame}

\item{\code{scores}}{list}
}}

