% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{CI}
\alias{CI}
\title{Critical index}
\usage{
CI(n, Sd, pcc_in, pcc_out)
}
\arguments{
\item{n}{sample size}

\item{Sd}{standard deviation}

\item{pcc_in}{pcc inside the group}

\item{pcc_out}{pcc outside the group}
}
\description{
Critical index
}
