% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DNBcompute}
\alias{DNBcompute}
\title{DNBcompute}
\usage{
DNBcompute(
  data,
  meta,
  diffgenes = NULL,
  allgenes = NULL,
  meta_levels = NULL,
  high_method = c("high_cv", "top_gene"),
  high_cutoff = 0.6,
  cutree_method = c("h", "k"),
  cutree_cutoff = 0.98,
  minModule = 7,
  maxModule = 60,
  quiet = FALSE,
  fastMode = FALSE,
  writefile = FALSE,
  cluster_fun = NULL,
  cluster_args = NULL,
  size_effect = TRUE
)
}
\arguments{
\item{data}{the gene expression matrix,
which can be a single-cell RNA-seq GEM with at least three group/clusters
or a matrix merging bulk GEMs from at least three different sample}

\item{meta}{a data.frame with rownames as cell-id as well as one column of group infomation}

\item{diffgenes}{which genes we're interested in, or no special ones (all, default)}

\item{allgenes}{the whole genes that ordered in advance by expression, or the rownames of GEM (default)}

\item{meta_levels}{the order of meta group, default ordered by decreasing if NULL}

\item{high_method}{the method to select genes for the first step, by either high_cv (default) or top_gene}

\item{high_cutoff}{the cutoff value corresponding to the high_method,

 with the range between 0 - 1(all) for high_cv and 1 - #allgenes(all) for top_gene

 or not to select highly variable genes but use all genes when -1}

\item{cutree_method}{the method to select numbers of tree (module) from hclust, 

 by either h (height, default) or k (number K)}

\item{cutree_cutoff}{the cutoff value corresponding to the cutree_method,

 with the range between 0-1 for h and a number greater than 0 for k}

\item{minModule}{the min number of genes of the module meeting requirements}

\item{maxModule}{the max number of genes of the module meeting requirements}

\item{quiet}{do not print output of process during calculation (against verbose), default FALSE}

\item{fastMode}{avoid using for loop, rathan apply-like function, default FALSE; if TRUE, quiet will be set as TRUE}

\item{writefile}{write results of each group into DNB_Module_information_xx.txt with tab delimiter, default FALSE}

\item{cluster_fun}{customized function that user design for clustering to find module, 

  default NULL (hierarchical by stats::hclust(d, method = "complete", members = NULL))
  
  This function should do function of clustering (e.g hclust + cutree), 
  
  with input that first arg "d" = distance matrix and output "named int vector". 
  
  If assigned, cutree_method and cutree_cutoff would be ignored}

\item{cluster_args}{a list of extra arguments to the cluster_fun call. 
The names attribute of args gives the argument names. (same to base::do.call(args))}

\item{size_effect}{whether consider the effect of sample size when compute CI of DNB, default TRUE}
}
\value{
S3:DNB_output
}
\description{
Compute the Dynamic Network Biomarkers(DNB) model
}
\details{
return a S3 object includes several S4 objects

(optional) write score results into DNB_score_matrix.txt
}
\examples{
data(data.example)
data(meta.example)
a <- DNBcompute(data.example, meta.example)
a

}
\author{
Kaiyu Wang, in ChenLab of CAS, Shanghai, China
}
