% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DNBcompute_custom}
\alias{DNBcompute_custom}
\title{DNBcompute_custom}
\usage{
DNBcompute_custom(data, meta, module_list, meta_levels = NULL)
}
\arguments{
\item{data}{the gene expression matrix,
which can be a single-cell RNA-seq GEM with at least three group/clusters
or a matrix merging bulk GEMs from at least three different sample}

\item{meta}{a data.frame with rownames as cell-id as well as one column of group infomation}

\item{module_list}{a customized list of module gene}

\item{meta_levels}{the order of meta group, default ordered by decreasing if NULL}
}
\value{
S3:DNB_output
}
\description{
Compute the Dynamic Network Biomarkers(DNB) model with customized Modules
}
\details{
return a S3 object includes several S4 objects
}
\examples{
data(data.example)
data(meta.example)
data(module_list.example)
b <- DNBcompute_custom(data.example, meta.example, module_list.example)
b
DNBplot(b, group = "USER_CUSTOMIZED", ranking = 1, show = TRUE, save_pdf = FALSE)
DNBplot(b, group = "USER_CUSTOMIZED", ranking = 2, show = TRUE, save_pdf = FALSE)

}
\author{
Kaiyu Wang, in ChenLab of CAS, Shanghai, China
}
