% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{DNBplot.DNB_output}
\alias{DNBplot.DNB_output}
\title{DNBplot for S3:DNB_output}
\usage{
\method{DNBplot}{DNB_output}(
  object,
  ranking = NULL,
  group = NULL,
  resource = NULL,
  show = TRUE,
  save_pdf = FALSE,
  file_prefix = NULL,
  meta_levels = NULL,
  ...
)
}
\arguments{
\item{object}{DNB_output}

\item{ranking}{ranking}

\item{group}{group}

\item{resource}{resource, exclusive with ranking & group}

\item{show}{whether to show}

\item{save_pdf}{whether to save pdf file}

\item{file_prefix}{the file prefix if save_pdf}

\item{meta_levels}{the order of meta-group in the plots, default levels(df_score$Names) if be null}

\item{...}{not use}
}
\description{
DNBplot for S3:DNB_output
}
