% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DNBplot}
\alias{DNBplot}
\title{DNBplot}
\usage{
DNBplot(
  object,
  ranking = NULL,
  group = NULL,
  show = TRUE,
  save_pdf = FALSE,
  file_prefix = NULL,
  meta_levels = NULL,
  ...
)
}
\arguments{
\item{object}{S3:DNB_output or df_score (output of ScoreExtract)}

\item{ranking}{for DNB_output}

\item{group}{for DNB_output}

\item{show}{whether to show}

\item{save_pdf}{whether to save pdf file}

\item{file_prefix}{the file prefix if save_pdf}

\item{meta_levels}{the order of meta-group in the plots, default levels(df_score$Names) if be null}

\item{...}{for future use}
}
\value{
plot or pdf
}
\description{
Visualize the DNB score information
}
\examples{
data(data.example)
data(meta.example)
data(diffgenes.example)
a <- DNBcompute(data.example, meta.example, diffgenes.example)
b <- DNBfilter(a, ntop = 5)
DNBplot(b, show = TRUE, save_pdf = FALSE)

}
\author{
Kaiyu Wang, in ChenLab of CAS, Shanghai, China
}
