% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{mynew.DNB_obj}
\alias{mynew.DNB_obj}
\title{initialize or generate the S4-DNB_obj}
\usage{
mynew.DNB_obj(data, ntop = NULL, ...)
}
\arguments{
\item{data}{raw data to compute DNB (matrix or data.frame)}

\item{ntop}{parameters for slot result, the number of top rank for modules}

\item{...}{parameters transmitting to function mynew.DNB_res() to initialize or generate the slot pre_result (S4-DNB_res)}
}
\value{
S4 object, DNB_obj
}
\description{
initialize or generate the S4-DNB_obj
}
