% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{ScoreExtract.DNB_output}
\alias{ScoreExtract.DNB_output}
\title{Extract the score information from object (S3:DNB_output)}
\usage{
\method{ScoreExtract}{DNB_output}(object, ranking = NULL, group = NULL, ...)
}
\arguments{
\item{object}{S3:DNB_output}

\item{ranking}{the ranking of exactly module}

\item{group}{which group to select module}

\item{...}{not use}
}
\value{
score data.frame
}
\description{
Extract the score information from object (S3:DNB_output)
}
