% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{singleDNB}
\alias{singleDNB}
\title{Compute data for the modules list, and return score matrix}
\usage{
singleDNB(data, module_list, cor_vec = NULL, size_effect = TRUE)
}
\arguments{
\item{data}{data}

\item{module_list}{modules list}

\item{cor_vec}{vector of correlation coefficient, when data has only one sample (col)}

\item{size_effect}{whether consider the effect of sample size when compute CI of DNB, default TRUE}
}
\value{
score matrix
}
\description{
Compute data for the modules list, and return score matrix
}
